requirements.txt
streamlit
requests
python-dotenv
pyyaml
bcrypt
